app/requirements.txt
dash==0.19.0  # The core dash backend
dash-renderer==0.11.1  # The dash front-end
dash-html-components==0.8.0  # HTML components
dash-core-components==0.15.2  # Supercharged components
plotly --upgrade  # Latest Plotly graphing library
python-firebase
sseclient-py